refactor flatten out into its own module
