+++ pipeline/test_beam_tables.py
@@ -24,12 +24,11 @@ from apache_beam.testing.test_pipeline import TestPipeline
 import apache_beam.testing.util as beam_test_util
 
 from pipeline import beam_tables
-from pipeline.metadata.blockpage import BlockpageMatcher
 from pipeline.metadata.fake_caida_ip_metadata import FakeCaidaIpMetadata
 from pipeline.metadata.maxmind import FakeMaxmindIpMetadata
+from pipeline.metadata import flatten
 
 
-# pylint: disable=too-many-lines
 class PipelineMainTest(unittest.TestCase):
   """Unit tests for beam pipeline steps."""
 
@@ -116,18 +115,6 @@ class PipelineMainTest(unittest.TestCase):
     full_name = runner._get_full_table_name('prod.echo_scan')
     self.assertEqual(full_name, 'firehook-censoredplanet:prod.echo_scan')
 
-  def test_source_from_filename(self) -> None:
-    """Test getting the data source name from the filename."""
-    self.assertEqual(
-        beam_tables._source_from_filename(
-            'gs://firehook-scans/echo/CP_Quack-echo-2020-08-23-06-01-02/results.json'
-        ), 'CP_Quack-echo-2020-08-23-06-01-02')
-
-    self.assertEqual(
-        beam_tables._source_from_filename(
-            'gs://firehook-scans/http/CP_Quack-http-2020-09-13-01-02-07/results.json'
-        ), 'CP_Quack-http-2020-09-13-01-02-07')
-
   def test_read_scan_text(self) -> None:  # pylint: disable=no-self-use
     """Test reading lines from compressed and uncompressed files"""
     p = TestPipeline()
@@ -167,413 +154,6 @@ class PipelineMainTest(unittest.TestCase):
         beam_tables._between_dates(filename, None, datetime.date(2020, 5, 10)))
     self.assertFalse(
         beam_tables._between_dates(filename, datetime.date(2020, 5, 12), None))
-
-  def test_parse_received_headers(self) -> None:
-    """Test parsing HTTP/S header fields into a flat format."""
-    headers = {
-        'Content-Language': ['en', 'fr'],
-        'Content-Type': ['text/html; charset=iso-8859-1']
-    }
-
-    flat_headers = beam_tables._parse_received_headers(headers)
-
-    expected_headers = [
-        'Content-Language: en', 'Content-Language: fr',
-        'Content-Type: text/html; charset=iso-8859-1'
-    ]
-
-    self.assertListEqual(flat_headers, expected_headers)
-
-  def test_parse_received_data_http(self) -> None:
-    """Test parsing sample HTTP data."""
-    # yapf: disable
-    received = {
-        'status_line': '403 Forbidden',
-        'headers': {},
-        'body': '<html><head><meta http-equiv="Content-Type" content="text/html; charset=windows-1256"><title>MNN3-1(1)</title></head><body><iframe src="http://10.10.34.35:80" style="width: 100%; height: 100%" scrolling="no" marginwidth="0" marginheight="0" frameborder="0" vspace="0" hspace="0"></iframe></body></html>\r\n\r\n'
-    }
-
-    expected = {
-        'received_status': '403 Forbidden',
-        'received_body': '<html><head><meta http-equiv="Content-Type" content="text/html; charset=windows-1256"><title>MNN3-1(1)</title></head><body><iframe src="http://10.10.34.35:80" style="width: 100%; height: 100%" scrolling="no" marginwidth="0" marginheight="0" frameborder="0" vspace="0" hspace="0"></iframe></body></html>\r\n\r\n',
-        'received_headers': [],
-        'blockpage': True,
-        'page_signature': 'b_nat_ir_national_1',
-    }
-    # yapf: enable
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    parsed = flattener._parse_received_data(received, True)
-    self.assertDictEqual(parsed, expected)
-
-  def test_parse_received_data_no_header_field(self) -> None:
-    """Test parsing reciveed HTTP/S data missing a header field."""
-    # yapf: disable
-    received = {
-        'status_line': '403 Forbidden',
-        'body': '<test-body>'
-        # No 'headers' field
-    }
-
-    expected = {
-        'received_status': '403 Forbidden',
-        'received_body': '<test-body>',
-        'received_headers': [],
-        'blockpage': None,
-        'page_signature': None,
-    }
-    # yapf: enable
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    parsed = flattener._parse_received_data(received, True)
-    self.assertDictEqual(parsed, expected)
-
-  def test_parse_received_data_https(self) -> None:
-    """Test parsing example HTTPS data."""
-    # yapf: disable
-    received = {
-        'status_line': '403 Forbidden',
-        'headers': {
-            'Content-Length': ['278'],
-            'Content-Type': ['text/html'],
-            'Date': ['Fri, 06 Nov 2020 20:24:19 GMT'],
-            'Expires': ['Fri, 06 Nov 2020 20:24:19 GMT'],
-            'Mime-Version': ['1.0'],
-            'Server': ['AkamaiGHost'],
-            'Set-Cookie': [
-                'bm_sz=6A1BDB4DFCA371F55C598A6D50C7DC3F~YAAQtTXdWKzJ+ZR1AQAA6zY7nwmc3d1xb2D5pqi3WHoMGfNsB8zB22LP5Kz/15sxdI3d3qznv4NzhGdb6CjijzFezAd18NREhybEvZMSZe2JHkjBjli/y1ZRMgC512ln7CCHURjS03UWDIzVrpwPV3Z/h/mq00NF2+LgHsDPelEZoArYVmEwH7OtE4zHAePErKw=; Domain=.discover.com; Path=/; Expires=Sat, 07 Nov 2020 00:24:19 GMT; Max-Age=14400; HttpOnly_abck=7A29878FA7120EC680C6E591A8FF3F5A~-1~YAAQtTXdWK3J+ZR1AQAA6zY7nwR93cThkIxWHn0icKtS6Wgb6NVHSQ80nZ6I2DzczA+1qn/0rXSGZUcFvW/+7tmDF0lHsieeRwnmIydhPELwAsNLjfBMF1lJm9Y7u4ppUpD4WtbRJ1g+Qhd9CLcelH3fQ8AVmJn/jRNN8WrisA8GKuUhpfjv9Gp1aGGqzv12H8u3Ogt/9oOv4Y8nKuS7CWipsFuPftCMeTBVbPn/JsV/NzttmkuFikLj8PwmpNecqlhaH1Ra32XDl/hVsCFWaPm4wdWO3d2WDK8Em0sHzklyTV4iFo6itVlCEHQ=~-1~-1~-1; Domain=.discover.com; Path=/; Expires=Sat, 06 Nov 2021 20:24:19 GMT; Max-Age=31536000; Secure'
-            ]
-        },
-        'body': '<HTML><HEAD>\n<TITLE>Access Denied</TITLE>\n</HEAD><BODY>\n<H1>Access Denied</H1>\n \nYou don\'t have permission to access "discover.com" on this server.<P>\nReference 18b535dd581604694259a71c660\n</BODY>\n</HTML>\n',
-        'tls': {
-            'version': 771,
-            'cipher_suite': 49199,
-            'cert': 'MIIG1DCCBbygAwIBAgIQBFzDKr18mq0F13LVYhA6FjANBgkqhkiG9w0BAQsFADB1MQswCQYDVQQGEwJVUzEVMBMGA1UEChMMRGlnaUNlcnQgSW5jMRkwFwYDVQQLExB3d3cuZGlnaWNlcnQuY29tMTQwMgYDVQQDEytEaWdpQ2VydCBTSEEyIEV4dGVuZGVkIFZhbGlkYXRpb24gU2VydmVyIENBMB4XDTIwMTAwNTAwMDAwMFoXDTIxMTAwNjEyMDAwMFowgdQxHTAbBgNVBA8MFFByaXZhdGUgT3JnYW5pemF0aW9uMRMwEQYLKwYBBAGCNzwCAQMTAkpQMRcwFQYDVQQFEw4wMTAwLTAxLTAwODgyNDELMAkGA1UEBhMCSlAxDjAMBgNVBAgTBVRva3lvMRMwEQYDVQQHEwpDaGl5b2RhLUt1MTkwNwYDVQQKEzBUb2tpbyBNYXJpbmUgYW5kIE5pY2hpZG8gRmlyZSBJbnN1cmFuY2UgQ28uIEx0ZC4xGDAWBgNVBAMTD3d3dy50YWJpa29yZS5qcDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAN+0RFcFIoCHvFTJs/+hexC5SxrKDAytiHNDzXLYaOFuf2LA+7UN33QE8dnINmV0ma7Udd1r8KmXJWJPeTxIJyskad8VNwx0oF00ENS56GYl/y37Y85DE5MQhaQwPEiyQL0TsrL/K2bNYjvEPklBVEOi1vtiOOTZWnUH86MxSe3PwmmXDaFgd3174Z8lEmi20Jl3++Tr/jNeBMw3Sg3KuLW8IUTl6+33mr3Z1u2u6yFN4d7mXlzyo0BxOwlJ1NwJbTzyFnBAfAZ2gJFVFQtuoWdgh9XIquhdFoxCfj/h9zxFK+64xJ+sXGSL5SiEZeBfmvG8SrW4OBSvHzyUSzJKCrsCAwEAAaOCAv4wggL6MB8GA1UdIwQYMBaAFD3TUKXWoK3u80pgCmXTIdT4+NYPMB0GA1UdDgQWBBQKix8NngHND9LiEWxMPAOBE6MwjDAnBgNVHREEIDAeggt0YWJpa29yZS5qcIIPd3d3LnRhYmlrb3JlLmpwMA4GA1UdDwEB/wQEAwIFoDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwdQYDVR0fBG4wbDA0oDKgMIYuaHR0cDovL2NybDMuZGlnaWNlcnQuY29tL3NoYTItZXYtc2VydmVyLWczLmNybDA0oDKgMIYuaHR0cDovL2NybDQuZGlnaWNlcnQuY29tL3NoYTItZXYtc2VydmVyLWczLmNybDBLBgNVHSAERDBCMDcGCWCGSAGG/WwCATAqMCgGCCsGAQUFBwIBFhxodHRwczovL3d3dy5kaWdpY2VydC5jb20vQ1BTMAcGBWeBDAEBMIGIBggrBgEFBQcBAQR8MHowJAYIKwYBBQUHMAGGGGh0dHA6Ly9vY3NwLmRpZ2ljZXJ0LmNvbTBSBggrBgEFBQcwAoZGaHR0cDovL2NhY2VydHMuZGlnaWNlcnQuY29tL0RpZ2lDZXJ0U0hBMkV4dGVuZGVkVmFsaWRhdGlvblNlcnZlckNBLmNydDAJBgNVHRMEAjAAMIIBBAYKKwYBBAHWeQIEAgSB9QSB8gDwAHYA9lyUL9F3MCIUVBgIMJRWjuNNExkzv98MLyALzE7xZOMAAAF093gqNAAABAMARzBFAiEAz0WGut1b8na4VKfulIqCPRbV+lv05YdPNT2xfWreNAYCIDU3JiavbsMjE/r0M9P2c7B07U72W4TK/PdlsKCg5t1PAHYAXNxDkv7mq0VEsV6a1FbmEDf71fpH3KFzlLJe5vbHDsoAAAF093gqgwAABAMARzBFAiApVQum+1q4C4drBI7t6aObwa5jtmWd/BHVTLPgcdhMXgIhAKv+7bC9X9wstKB0OGQbVVX/qsJ5fzf4Y8zNUaklAQiKMA0GCSqGSIb3DQEBCwUAA4IBAQAD02pESpGPgJSMTpFVm4VRufgwW95fxA/sch63U94owcOmNtrniSoOr8QwLMAVta6VFU6wddbTBd4vz8zauo4R6uAeFaiUBaFaKb5V2bONGclfjTZ7nsDxsowLracGrRx/rQjjovRo2656g5Iu898WIfADxIvsGc5CICGqLB9GvofVWNNb/DoOXf/vLQJj9m5+ZCi0CrIdh31IB/acHsQ8jWr4VlqPGiz2PIdKjBLuI9ckFbMQ/9DCTWfuJhSfwA3kk2EeUa6WlRrjDhJLasjrEmQiSIf3oywdsPspSYOkT91TFUvzjOmK/yZeApxPJmDvjxpum5GZYnn6QthKxMzL'
-        }
-    }
-
-    expected = {
-        'received_status': '403 Forbidden',
-        'received_body': '<HTML><HEAD>\n<TITLE>Access Denied</TITLE>\n</HEAD><BODY>\n<H1>Access Denied</H1>\n \nYou don\'t have permission to access "discover.com" on this server.<P>\nReference 18b535dd581604694259a71c660\n</BODY>\n</HTML>\n',
-        'received_tls_version': 771,
-        'received_tls_cipher_suite': 49199,
-        'received_tls_cert': 'MIIG1DCCBbygAwIBAgIQBFzDKr18mq0F13LVYhA6FjANBgkqhkiG9w0BAQsFADB1MQswCQYDVQQGEwJVUzEVMBMGA1UEChMMRGlnaUNlcnQgSW5jMRkwFwYDVQQLExB3d3cuZGlnaWNlcnQuY29tMTQwMgYDVQQDEytEaWdpQ2VydCBTSEEyIEV4dGVuZGVkIFZhbGlkYXRpb24gU2VydmVyIENBMB4XDTIwMTAwNTAwMDAwMFoXDTIxMTAwNjEyMDAwMFowgdQxHTAbBgNVBA8MFFByaXZhdGUgT3JnYW5pemF0aW9uMRMwEQYLKwYBBAGCNzwCAQMTAkpQMRcwFQYDVQQFEw4wMTAwLTAxLTAwODgyNDELMAkGA1UEBhMCSlAxDjAMBgNVBAgTBVRva3lvMRMwEQYDVQQHEwpDaGl5b2RhLUt1MTkwNwYDVQQKEzBUb2tpbyBNYXJpbmUgYW5kIE5pY2hpZG8gRmlyZSBJbnN1cmFuY2UgQ28uIEx0ZC4xGDAWBgNVBAMTD3d3dy50YWJpa29yZS5qcDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAN+0RFcFIoCHvFTJs/+hexC5SxrKDAytiHNDzXLYaOFuf2LA+7UN33QE8dnINmV0ma7Udd1r8KmXJWJPeTxIJyskad8VNwx0oF00ENS56GYl/y37Y85DE5MQhaQwPEiyQL0TsrL/K2bNYjvEPklBVEOi1vtiOOTZWnUH86MxSe3PwmmXDaFgd3174Z8lEmi20Jl3++Tr/jNeBMw3Sg3KuLW8IUTl6+33mr3Z1u2u6yFN4d7mXlzyo0BxOwlJ1NwJbTzyFnBAfAZ2gJFVFQtuoWdgh9XIquhdFoxCfj/h9zxFK+64xJ+sXGSL5SiEZeBfmvG8SrW4OBSvHzyUSzJKCrsCAwEAAaOCAv4wggL6MB8GA1UdIwQYMBaAFD3TUKXWoK3u80pgCmXTIdT4+NYPMB0GA1UdDgQWBBQKix8NngHND9LiEWxMPAOBE6MwjDAnBgNVHREEIDAeggt0YWJpa29yZS5qcIIPd3d3LnRhYmlrb3JlLmpwMA4GA1UdDwEB/wQEAwIFoDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwdQYDVR0fBG4wbDA0oDKgMIYuaHR0cDovL2NybDMuZGlnaWNlcnQuY29tL3NoYTItZXYtc2VydmVyLWczLmNybDA0oDKgMIYuaHR0cDovL2NybDQuZGlnaWNlcnQuY29tL3NoYTItZXYtc2VydmVyLWczLmNybDBLBgNVHSAERDBCMDcGCWCGSAGG/WwCATAqMCgGCCsGAQUFBwIBFhxodHRwczovL3d3dy5kaWdpY2VydC5jb20vQ1BTMAcGBWeBDAEBMIGIBggrBgEFBQcBAQR8MHowJAYIKwYBBQUHMAGGGGh0dHA6Ly9vY3NwLmRpZ2ljZXJ0LmNvbTBSBggrBgEFBQcwAoZGaHR0cDovL2NhY2VydHMuZGlnaWNlcnQuY29tL0RpZ2lDZXJ0U0hBMkV4dGVuZGVkVmFsaWRhdGlvblNlcnZlckNBLmNydDAJBgNVHRMEAjAAMIIBBAYKKwYBBAHWeQIEAgSB9QSB8gDwAHYA9lyUL9F3MCIUVBgIMJRWjuNNExkzv98MLyALzE7xZOMAAAF093gqNAAABAMARzBFAiEAz0WGut1b8na4VKfulIqCPRbV+lv05YdPNT2xfWreNAYCIDU3JiavbsMjE/r0M9P2c7B07U72W4TK/PdlsKCg5t1PAHYAXNxDkv7mq0VEsV6a1FbmEDf71fpH3KFzlLJe5vbHDsoAAAF093gqgwAABAMARzBFAiApVQum+1q4C4drBI7t6aObwa5jtmWd/BHVTLPgcdhMXgIhAKv+7bC9X9wstKB0OGQbVVX/qsJ5fzf4Y8zNUaklAQiKMA0GCSqGSIb3DQEBCwUAA4IBAQAD02pESpGPgJSMTpFVm4VRufgwW95fxA/sch63U94owcOmNtrniSoOr8QwLMAVta6VFU6wddbTBd4vz8zauo4R6uAeFaiUBaFaKb5V2bONGclfjTZ7nsDxsowLracGrRx/rQjjovRo2656g5Iu898WIfADxIvsGc5CICGqLB9GvofVWNNb/DoOXf/vLQJj9m5+ZCi0CrIdh31IB/acHsQ8jWr4VlqPGiz2PIdKjBLuI9ckFbMQ/9DCTWfuJhSfwA3kk2EeUa6WlRrjDhJLasjrEmQiSIf3oywdsPspSYOkT91TFUvzjOmK/yZeApxPJmDvjxpum5GZYnn6QthKxMzL',
-        'received_headers': [
-            'Content-Length: 278',
-            'Content-Type: text/html',
-            'Date: Fri, 06 Nov 2020 20:24:19 GMT',
-            'Expires: Fri, 06 Nov 2020 20:24:19 GMT',
-            'Mime-Version: 1.0',
-            'Server: AkamaiGHost',
-            'Set-Cookie: bm_sz=6A1BDB4DFCA371F55C598A6D50C7DC3F~YAAQtTXdWKzJ+ZR1AQAA6zY7nwmc3d1xb2D5pqi3WHoMGfNsB8zB22LP5Kz/15sxdI3d3qznv4NzhGdb6CjijzFezAd18NREhybEvZMSZe2JHkjBjli/y1ZRMgC512ln7CCHURjS03UWDIzVrpwPV3Z/h/mq00NF2+LgHsDPelEZoArYVmEwH7OtE4zHAePErKw=; Domain=.discover.com; Path=/; Expires=Sat, 07 Nov 2020 00:24:19 GMT; Max-Age=14400; HttpOnly_abck=7A29878FA7120EC680C6E591A8FF3F5A~-1~YAAQtTXdWK3J+ZR1AQAA6zY7nwR93cThkIxWHn0icKtS6Wgb6NVHSQ80nZ6I2DzczA+1qn/0rXSGZUcFvW/+7tmDF0lHsieeRwnmIydhPELwAsNLjfBMF1lJm9Y7u4ppUpD4WtbRJ1g+Qhd9CLcelH3fQ8AVmJn/jRNN8WrisA8GKuUhpfjv9Gp1aGGqzv12H8u3Ogt/9oOv4Y8nKuS7CWipsFuPftCMeTBVbPn/JsV/NzttmkuFikLj8PwmpNecqlhaH1Ra32XDl/hVsCFWaPm4wdWO3d2WDK8Em0sHzklyTV4iFo6itVlCEHQ=~-1~-1~-1; Domain=.discover.com; Path=/; Expires=Sat, 06 Nov 2021 20:24:19 GMT; Max-Age=31536000; Secure'
-        ],
-        'blockpage': False,
-        'page_signature': 'x_on_this_server',
-    }
-    # yapf: enable
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    parsed = flattener._parse_received_data(received, True)
-    self.assertDictEqual(parsed, expected)
-
-  def test_flattenmeasurement_echo(self) -> None:
-    """Test parsing an example Echo measurement."""
-    line = """{
-      "Server":"1.2.3.4",
-      "Keyword":"www.example.com",
-      "Retries":1,
-      "Results":[
-        {
-          "Sent":"GET / HTTP/1.1 Host: www.example.com",
-          "Received":"HTTP/1.1 403 Forbidden",
-          "Success":false,
-          "Error":"Incorrect echo response",
-          "StartTime":"2020-09-20T07:45:09.643770291-04:00",
-          "EndTime":"2020-09-20T07:45:10.088851843-04:00"
-        },
-        {
-          "Sent":"GET / HTTP/1.1 Host: www.example.com",
-          "Received": "HTTP/1.1 503 Service Unavailable",
-          "Success":false,
-          "Error":"Incorrect echo response",
-          "StartTime":"2020-09-20T07:45:16.170427683-04:00",
-          "EndTime":"2020-09-20T07:45:16.662093893-04:00"
-        }
-      ],
-      "Blocked":true,
-      "FailSanity":false,
-      "StatefulBlock":false
-    }"""
-
-    expected_rows: List[beam_tables.Row] = [{
-        'domain': 'www.example.com',
-        'ip': '1.2.3.4',
-        'date': '2020-09-20',
-        'start_time': '2020-09-20T07:45:09.643770291-04:00',
-        'end_time': '2020-09-20T07:45:10.088851843-04:00',
-        'retries': 1,
-        'sent': 'GET / HTTP/1.1 Host: www.example.com',
-        'received_status': 'HTTP/1.1 403 Forbidden',
-        'error': 'Incorrect echo response',
-        'blocked': True,
-        'success': False,
-        'fail_sanity': False,
-        'stateful_block': False,
-        'measurement_id': '',
-        'source': 'CP_Quack-echo-2020-08-23-06-01-02',
-    }, {
-        'domain': 'www.example.com',
-        'ip': '1.2.3.4',
-        'date': '2020-09-20',
-        'start_time': '2020-09-20T07:45:16.170427683-04:00',
-        'end_time': '2020-09-20T07:45:16.662093893-04:00',
-        'retries': 1,
-        'sent': 'GET / HTTP/1.1 Host: www.example.com',
-        'received_status': 'HTTP/1.1 503 Service Unavailable',
-        'error': 'Incorrect echo response',
-        'blocked': True,
-        'success': False,
-        'fail_sanity': False,
-        'stateful_block': False,
-        'measurement_id': '',
-        'source': 'CP_Quack-echo-2020-08-23-06-01-02',
-    }]
-
-    filename = 'gs://firehook-scans/echo/CP_Quack-echo-2020-08-23-06-01-02/results.json'
-
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    rows = list(flattener.process((filename, line)))
-    self.assertEqual(len(rows), 2)
-
-    # Measurement ids should be the same
-    self.assertEqual(rows[0]['measurement_id'], rows[1]['measurement_id'])
-    # But they're randomly generated,
-    # so we can't test them against the full expected rows.
-    rows[0]['measurement_id'] = ''
-    rows[1]['measurement_id'] = ''
-
-    self.assertListEqual(rows, expected_rows)
-
-  def test_flattenmeasurement_http_success(self) -> None:
-    """Test parsing an example successful HTTP measurement
-
-    Not all measurements recieve any data/errors,
-    in that case the received_ and error fields should not exist
-    and will end up Null in bigquery.
-    """
-
-    line = """{
-      "Server":"170.248.33.11",
-      "Keyword":"scribd.com",
-      "Retries":0,
-      "Results":[
-        {
-         "Sent":"scribd.com",
-         "Success":true,
-         "StartTime":"2020-11-09T01:10:47.826486107-05:00",
-         "EndTime":"2020-11-09T01:10:47.84869292-05:00"
-        }
-      ],
-      "Blocked":false,
-      "FailSanity":false,
-      "StatefulBlock":false
-    }"""
-
-    expected_row = {
-        'domain': 'scribd.com',
-        'ip': '170.248.33.11',
-        'date': '2020-11-09',
-        'start_time': '2020-11-09T01:10:47.826486107-05:00',
-        'end_time': '2020-11-09T01:10:47.84869292-05:00',
-        'retries': 0,
-        'sent': 'scribd.com',
-        'blocked': False,
-        'success': True,
-        'fail_sanity': False,
-        'stateful_block': False,
-        'measurement_id': '',
-        'source': 'CP_Quack-http-2020-11-09-01-02-08',
-    }
-
-    filename = 'gs://firehook-scans/http/CP_Quack-http-2020-11-09-01-02-08/results.json'
-
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    row = list(flattener.process((filename, line)))[0]
-    # We can't test the measurement id because it's random
-    row['measurement_id'] = ''
-    self.assertEqual(row, expected_row)
-
-  def test_flattenmeasurement_http(self) -> None:
-    """Test parsing an unsuccessful HTTP measurement."""
-    line = """{
-      "Server":"184.50.171.225",
-      "Keyword":"www.csmonitor.com",
-      "Retries":0,
-      "Results":[
-        {
-         "Sent":"www.csmonitor.com",
-         "Received":{
-            "status_line":"301 Moved Permanently",
-            "headers":{
-               "Content-Length":["0"],
-               "Date":["Sun, 13 Sep 2020 05:10:58 GMT"],
-               "Location":["https://www.csmonitor.com/"],
-               "Server":["HTTP Proxy/1.0"]
-            },
-            "body":"test body"
-         },
-         "Success":false,
-         "Error":"Incorrect web response: status lines don't match",
-         "StartTime":"2020-09-13T01:10:57.499263112-04:00",
-         "EndTime":"2020-09-13T01:10:58.077524926-04:00"
-        }
-      ],
-      "Blocked":true,
-      "FailSanity":false,
-      "StatefulBlock":false
-    }"""
-
-    expected_row = {
-        'domain': 'www.csmonitor.com',
-        'ip': '184.50.171.225',
-        'date': '2020-09-13',
-        'start_time': '2020-09-13T01:10:57.499263112-04:00',
-        'end_time': '2020-09-13T01:10:58.077524926-04:00',
-        'retries': 0,
-        'sent': 'www.csmonitor.com',
-        'received_status': '301 Moved Permanently',
-        'received_body': 'test body',
-        'received_headers': [
-            'Content-Length: 0',
-            'Date: Sun, 13 Sep 2020 05:10:58 GMT',
-            'Location: https://www.csmonitor.com/',
-            'Server: HTTP Proxy/1.0',
-        ],
-        'blockpage': None,
-        'page_signature': None,
-        'error': 'Incorrect web response: status lines don\'t match',
-        'blocked': True,
-        'success': False,
-        'fail_sanity': False,
-        'stateful_block': False,
-        'measurement_id': '',
-        'source': 'CP_Quack-http-2020-09-13-01-02-07',
-    }
-    filename = 'gs://firehook-scans/http/CP_Quack-http-2020-09-13-01-02-07/results.json'
-
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    row = list(flattener.process((filename, line)))[0]
-    # We can't test the measurement id because it's random
-    row['measurement_id'] = ''
-    self.assertEqual(row, expected_row)
-
-  def test_flattenmeasurement_https(self) -> None:
-    """Test parsing an unsuccessful HTTPS measurement."""
-    # yapf: disable
-    line = """{
-      "Server":"213.175.166.157",
-      "Keyword":"www.arabhra.org",
-      "Retries":2,
-      "Results":[
-        {
-         "Sent":"www.arabhra.org",
-         "Received":{
-           "status_line":"302 Found",
-           "headers":{
-             "Content-Language":["en"],
-             "Content-Type":["text/html; charset=iso-8859-1"],
-             "Date":["Fri, 06 Nov 2020 20:24:21 GMT"],
-             "Location":[
-               "https://jobs.bankaudi.com.lb/OA_HTML/IrcVisitor.jsp"
-             ],
-             "Set-Cookie":[
-               "BIGipServer~IFRMS-WEB~IFRMS-OHS-HTTPS=rd7o00000000000000000000ffffc0a8fde4o4443; expires=Fri, 06-Nov-2020 21:24:21 GMT; path=/; Httponly; Secure",
-               "TS016c74f4=01671efb9a1a400535e215d6f76498a5887425fed793ca942baa75f16076e60e1350988222922fa06fc16f53ef016d9ecd38535fcabf14861525811a7c3459e91086df326f; Path=/"
-             ],
-             "X-Frame-Options":["SAMEORIGIN"]
-            },
-            "body":"\\u003c!DOCTYPE HTML PUBLIC \\\"-//IETF//DTD HTML 2.0//EN\\\"\\u003e\\n\\u003cHTML\\u003e\\u003cHEAD\\u003e\\n\\u003cTITLE\\u003e302 Found\\u003c/TITLE\\u003e\\n\\u003c/HEAD\\u003e\\u003cBODY\\u003e\\n\\u003cH1\\u003eFound\\u003c/H1\\u003e\\nThe document has moved \\u003cA HREF=\\\"https://jobs.bankaudi.com.lb/OA_HTML/IrcVisitor.jsp\\\"\\u003ehere\\u003c/A\\u003e.\\u003cP\\u003e\\n\\u003c/BODY\\u003e\\u003c/HTML\\u003e\\n",
-            "tls":{
-              "version":771,
-              "cipher_suite":49199,
-              "cert":"MIIHLzCCBhegAwIBAgIQDCECYKFMPekAAAAAVNFY9jANBgkqhkiG9w0BAQsFADCBujELMAkGA1UEBhMCVVMxFjAUBgNVBAoTDUVudHJ1c3QsIEluYy4xKDAmBgNVBAsTH1NlZSB3d3cuZW50cnVzdC5uZXQvbGVnYWwtdGVybXMxOTA3BgNVBAsTMChjKSAyMDE0IEVudHJ1c3QsIEluYy4gLSBmb3IgYXV0aG9yaXplZCB1c2Ugb25seTEuMCwGA1UEAxMlRW50cnVzdCBDZXJ0aWZpY2F0aW9uIEF1dGhvcml0eSAtIEwxTTAeFw0yMDA1MTIxMTIzMDNaFw0yMTA1MTIxMTUzMDJaMIHCMQswCQYDVQQGEwJMQjEPMA0GA1UEBxMGQmVpcnV0MRMwEQYLKwYBBAGCNzwCAQMTAkxCMRcwFQYLKwYBBAGCNzwCAQETBkJlaXJ1dDEWMBQGA1UEChMNQmFuayBBdWRpIFNBTDEdMBsGA1UEDxMUUHJpdmF0ZSBPcmdhbml6YXRpb24xDjAMBgNVBAsTBUJBU0FMMQ4wDAYDVQQFEwUxMTM0NzEdMBsGA1UEAxMUam9icy5iYW5rYXVkaS5jb20ubGIwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQC37LFk2A2Q8xxahyjhOkul8O9Nv5FFp0NkL4qIy2fTUbsz1uWOqQKo0jDS6Inwtb+i84//znY7ed7Uu5LfbPk0Biefkl4ke0d9LZ3fu7y0iQWWUqKGn4YAPDGv3R0y/47XlhHhDaR+D0z7SbmYHx2NQI7fj6iEfEB90PvPhrdDEKHypNoXa5PwOuGSoU0l+yGmuvF5N7/hr82y987pLRjMdJaszs5EM//C+eiyL9mTA8gvOOf3ZHYQ4ITsJpA9I2Q0E6fDQhGS8SDW2ktdZ7z2TIOQsyMuXJKbBeXCgKyjnaX5UWDis8Hpj43CI8Kge32qsqaTKbjf3Mb66nqHrwSdAgMBAAGjggMlMIIDITA5BgNVHREEMjAwghRqb2JzLmJhbmthdWRpLmNvbS5sYoIYd3d3LmpvYnMuYmFua2F1ZGkuY29tLmxiMIIBfQYKKwYBBAHWeQIEAgSCAW0EggFpAWcAdQBVgdTCFpA2AUrqC5tXPFPwwOQ4eHAlCBcvo6odBxPTDAAAAXIIuy40AAAEAwBGMEQCIEByP85HYDmBb/4WK0B6s5L66Owim+Hzf3jiPYvzhw5eAiBsT1ZEn5PuJfBZ9a9Y/TzJ8K9Qx+3+pyJATsPglI4z3AB2AJQgvB6O1Y1siHMfgosiLA3R2k1ebE+UPWHbTi9YTaLCAAABcgi7LlQAAAQDAEcwRQIgOgyG1ORFwA+sDB3cD4fCu25ahSyMi/4d+xvrP+STJxgCIQDXm1WBzc+gQlU/PhpVti+e4j+2MouWIBBvjw3k0/HTtgB2APZclC/RdzAiFFQYCDCUVo7jTRMZM7/fDC8gC8xO8WTjAAABcgi7LqAAAAQDAEcwRQIgaiMkFpZwGZ5Iac/cfTL8v6TbPHUIeSVjTnB1Z2m9gsoCIQCJr+wqJ0UF+FYhxq9ChDfn1Ukg3uVQePrv4WoWNYjOZzAOBgNVHQ8BAf8EBAMCBaAwHQYDVR0lBBYwFAYIKwYBBQUHAwEGCCsGAQUFBwMCMGgGCCsGAQUFBwEBBFwwWjAjBggrBgEFBQcwAYYXaHR0cDovL29jc3AuZW50cnVzdC5uZXQwMwYIKwYBBQUHMAKGJ2h0dHA6Ly9haWEuZW50cnVzdC5uZXQvbDFtLWNoYWluMjU2LmNlcjAzBgNVHR8ELDAqMCigJqAkhiJodHRwOi8vY3JsLmVudHJ1c3QubmV0L2xldmVsMW0uY3JsMEoGA1UdIARDMEEwNgYKYIZIAYb6bAoBAjAoMCYGCCsGAQUFBwIBFhpodHRwOi8vd3d3LmVudHJ1c3QubmV0L3JwYTAHBgVngQwBATAfBgNVHSMEGDAWgBTD99C1KjCtrw2RIXA5VN28iXDHOjAdBgNVHQ4EFgQUt5uewiz6lN1FGnoOCX/soGsCwoIwCQYDVR0TBAIwADANBgkqhkiG9w0BAQsFAAOCAQEArlnXiyOefAVaQd0jfxtGwzAed4c8EidlBaoebJACR4zlAIFG0r0pXbdHkLZnCkMCL7XvoV+Y27c1I/Tfcket6qr4gDuKKnbUZIdgg8LGU2OklVEfLv1LJi3+tRuGGCfKpzHWoL1FW+3T6YEETGeb1qZrGBE7Its/4WfVAwaBHynSrvdjeQTYuYP8XsvehhfI5PNQbfV3KIH+sOF7sg80C2sIEyxwD+VEfRGeV6nEhJGJdlibAWfNOwQAyRQcGoiVIdLoa9um9UAUugjktJJ/Dk74YyxIf3aX1yjqTANVIuBgSotC8FvUNTmAALL7Ug8fqvJ9sPQhxIataKh/JdrDCQ=="
-            }
-          },
-          "Success":false,
-          "Error":"Incorrect web response: status lines don't match",
-          "StartTime":"2020-11-06T15:24:21.124508839-05:00",
-          "EndTime":"2020-11-06T15:24:21.812075476-05:00"
-        }
-      ],
-      "Blocked":false,
-      "FailSanity":false,
-      "StatefulBlock":false
-    }
-    """
-    # yapf: enable
-
-    filename = 'gs://firehook-scans/http/CP_Quack-https-2020-11-06-15-15-31/results.json'
-
-    # yapf: disable
-    expected_row: beam_tables.Row = {
-        'domain': 'www.arabhra.org',
-        'ip': '213.175.166.157',
-        'date': '2020-11-06',
-        'start_time': '2020-11-06T15:24:21.124508839-05:00',
-        'end_time': '2020-11-06T15:24:21.812075476-05:00',
-        'retries': 2,
-        'sent': 'www.arabhra.org',
-        'received_status': '302 Found',
-        # The received_body field in the json has a lot of unicode escapes
-        # but the interpreted string in the output should not.
-        'received_body': '<!DOCTYPE HTML PUBLIC \"-//IETF//DTD HTML 2.0//EN\">\n<HTML><HEAD>\n<TITLE>302 Found</TITLE>\n</HEAD><BODY>\n<H1>Found</H1>\nThe document has moved <A HREF=\"https://jobs.bankaudi.com.lb/OA_HTML/IrcVisitor.jsp\">here</A>.<P>\n</BODY></HTML>\n',
-        'received_tls_version': 771,
-        'received_tls_cipher_suite': 49199,
-        'received_tls_cert': 'MIIHLzCCBhegAwIBAgIQDCECYKFMPekAAAAAVNFY9jANBgkqhkiG9w0BAQsFADCBujELMAkGA1UEBhMCVVMxFjAUBgNVBAoTDUVudHJ1c3QsIEluYy4xKDAmBgNVBAsTH1NlZSB3d3cuZW50cnVzdC5uZXQvbGVnYWwtdGVybXMxOTA3BgNVBAsTMChjKSAyMDE0IEVudHJ1c3QsIEluYy4gLSBmb3IgYXV0aG9yaXplZCB1c2Ugb25seTEuMCwGA1UEAxMlRW50cnVzdCBDZXJ0aWZpY2F0aW9uIEF1dGhvcml0eSAtIEwxTTAeFw0yMDA1MTIxMTIzMDNaFw0yMTA1MTIxMTUzMDJaMIHCMQswCQYDVQQGEwJMQjEPMA0GA1UEBxMGQmVpcnV0MRMwEQYLKwYBBAGCNzwCAQMTAkxCMRcwFQYLKwYBBAGCNzwCAQETBkJlaXJ1dDEWMBQGA1UEChMNQmFuayBBdWRpIFNBTDEdMBsGA1UEDxMUUHJpdmF0ZSBPcmdhbml6YXRpb24xDjAMBgNVBAsTBUJBU0FMMQ4wDAYDVQQFEwUxMTM0NzEdMBsGA1UEAxMUam9icy5iYW5rYXVkaS5jb20ubGIwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQC37LFk2A2Q8xxahyjhOkul8O9Nv5FFp0NkL4qIy2fTUbsz1uWOqQKo0jDS6Inwtb+i84//znY7ed7Uu5LfbPk0Biefkl4ke0d9LZ3fu7y0iQWWUqKGn4YAPDGv3R0y/47XlhHhDaR+D0z7SbmYHx2NQI7fj6iEfEB90PvPhrdDEKHypNoXa5PwOuGSoU0l+yGmuvF5N7/hr82y987pLRjMdJaszs5EM//C+eiyL9mTA8gvOOf3ZHYQ4ITsJpA9I2Q0E6fDQhGS8SDW2ktdZ7z2TIOQsyMuXJKbBeXCgKyjnaX5UWDis8Hpj43CI8Kge32qsqaTKbjf3Mb66nqHrwSdAgMBAAGjggMlMIIDITA5BgNVHREEMjAwghRqb2JzLmJhbmthdWRpLmNvbS5sYoIYd3d3LmpvYnMuYmFua2F1ZGkuY29tLmxiMIIBfQYKKwYBBAHWeQIEAgSCAW0EggFpAWcAdQBVgdTCFpA2AUrqC5tXPFPwwOQ4eHAlCBcvo6odBxPTDAAAAXIIuy40AAAEAwBGMEQCIEByP85HYDmBb/4WK0B6s5L66Owim+Hzf3jiPYvzhw5eAiBsT1ZEn5PuJfBZ9a9Y/TzJ8K9Qx+3+pyJATsPglI4z3AB2AJQgvB6O1Y1siHMfgosiLA3R2k1ebE+UPWHbTi9YTaLCAAABcgi7LlQAAAQDAEcwRQIgOgyG1ORFwA+sDB3cD4fCu25ahSyMi/4d+xvrP+STJxgCIQDXm1WBzc+gQlU/PhpVti+e4j+2MouWIBBvjw3k0/HTtgB2APZclC/RdzAiFFQYCDCUVo7jTRMZM7/fDC8gC8xO8WTjAAABcgi7LqAAAAQDAEcwRQIgaiMkFpZwGZ5Iac/cfTL8v6TbPHUIeSVjTnB1Z2m9gsoCIQCJr+wqJ0UF+FYhxq9ChDfn1Ukg3uVQePrv4WoWNYjOZzAOBgNVHQ8BAf8EBAMCBaAwHQYDVR0lBBYwFAYIKwYBBQUHAwEGCCsGAQUFBwMCMGgGCCsGAQUFBwEBBFwwWjAjBggrBgEFBQcwAYYXaHR0cDovL29jc3AuZW50cnVzdC5uZXQwMwYIKwYBBQUHMAKGJ2h0dHA6Ly9haWEuZW50cnVzdC5uZXQvbDFtLWNoYWluMjU2LmNlcjAzBgNVHR8ELDAqMCigJqAkhiJodHRwOi8vY3JsLmVudHJ1c3QubmV0L2xldmVsMW0uY3JsMEoGA1UdIARDMEEwNgYKYIZIAYb6bAoBAjAoMCYGCCsGAQUFBwIBFhpodHRwOi8vd3d3LmVudHJ1c3QubmV0L3JwYTAHBgVngQwBATAfBgNVHSMEGDAWgBTD99C1KjCtrw2RIXA5VN28iXDHOjAdBgNVHQ4EFgQUt5uewiz6lN1FGnoOCX/soGsCwoIwCQYDVR0TBAIwADANBgkqhkiG9w0BAQsFAAOCAQEArlnXiyOefAVaQd0jfxtGwzAed4c8EidlBaoebJACR4zlAIFG0r0pXbdHkLZnCkMCL7XvoV+Y27c1I/Tfcket6qr4gDuKKnbUZIdgg8LGU2OklVEfLv1LJi3+tRuGGCfKpzHWoL1FW+3T6YEETGeb1qZrGBE7Its/4WfVAwaBHynSrvdjeQTYuYP8XsvehhfI5PNQbfV3KIH+sOF7sg80C2sIEyxwD+VEfRGeV6nEhJGJdlibAWfNOwQAyRQcGoiVIdLoa9um9UAUugjktJJ/Dk74YyxIf3aX1yjqTANVIuBgSotC8FvUNTmAALL7Ug8fqvJ9sPQhxIataKh/JdrDCQ==',
-        'received_headers': [
-            'Content-Language: en',
-            'Content-Type: text/html; charset=iso-8859-1',
-            'Date: Fri, 06 Nov 2020 20:24:21 GMT',
-            'Location: https://jobs.bankaudi.com.lb/OA_HTML/IrcVisitor.jsp',
-            'Set-Cookie: BIGipServer~IFRMS-WEB~IFRMS-OHS-HTTPS=rd7o00000000000000000000ffffc0a8fde4o4443; expires=Fri, 06-Nov-2020 21:24:21 GMT; path=/; Httponly; Secure',
-            'Set-Cookie: TS016c74f4=01671efb9a1a400535e215d6f76498a5887425fed793ca942baa75f16076e60e1350988222922fa06fc16f53ef016d9ecd38535fcabf14861525811a7c3459e91086df326f; Path=/',
-            'X-Frame-Options: SAMEORIGIN',
-        ],
-        'error': 'Incorrect web response: status lines don\'t match',
-        'blocked': False,
-        'success': False,
-        'fail_sanity': False,
-        'stateful_block': False,
-        'measurement_id': '',
-        'source': 'CP_Quack-https-2020-11-06-15-15-31',
-    }
-    # yapf: enable
-
-    flattener = beam_tables.FlattenMeasurement()
-    flattener.setup()
-    row = list(flattener.process((filename, line)))[0]
-    # We can't test the measurement id because it's random
-    row['measurement_id'] = ''
-
-    self.assertEqual(row, expected_row)
-
-  def test_flattenmeasurement_invalid_json(self) -> None:
-    """Test logging an error when parsing invalid JSON."""
-    line = 'invalid json'
-
-    with self.assertLogs(level='WARNING') as cm:
-      flattener = beam_tables.FlattenMeasurement()
-      flattener.setup()
-      rows = list(flattener.process(('test_filename.json', line)))
-      self.assertEqual(
-          cm.output[0], 'WARNING:root:JSONDecodeError: '
-          'Expecting value: line 1 column 1 (char 0)\n'
-          'Filename: test_filename.json\ninvalid json\n')
-
-    self.assertEqual(len(rows), 0)
 
   def test_add_metadata(self) -> None:  # pylint: disable=no-self-use
     """Test adding IP metadata to mesurements."""
@@ -756,209 +336,6 @@ class PipelineMainTest(unittest.TestCase):
 
     rows_with_metadata = list(beam_tables._merge_metadata_with_rows(key, value))
     self.assertListEqual(rows_with_metadata, expected_rows)
-
-  def test_blockpage_matching(self) -> None:
-    """Test blockpage matching for data with detected anomalies."""
-    matcher = BlockpageMatcher()
-
-    # yapf: disable
-    pages = [
-      '\u003cHTML\u003e\u003cHEAD\u003e\u003cTITLE\u003eError\u003c/TITLE\u003e\u003c/HEAD\u003e\u003cBODY\u003e\nAn error occurred while processing your request.\u003cp\u003e\nReference\u0026#32;\u0026#35;218\u0026#46;7714d23c\u0026#46;1568241844\u0026#46;522b6\n\u003c/BODY\u003e\u003c/HTML\u003e\n',
-      'URL blocked by SpIDer Gate',
-      '\u003cHTML\u003e\u003cHEAD\u003e\u003cTITLE\u003eA Website\u003c/TITLE\u003e\u003c/HEAD\u003e\u003cBODY\u003e\nThis site is not blocked!\u003cp\u003e\n\u003c/BODY\u003e\u003c/HTML\u003e\n',
-      'HTTP/1.1 302 Moved Temporarily\r\nLocation: iterika.ru/blocked.html?UrlRedir=http%3A%2F%2Fblackberry.com%2f\r\nContent-Length: 0\r\nCache-Control: max-age=0, no-cache, no-store, must-revalidate\r\nPragma: no-cache\r\nConnection: close\r\n\r\n',
-      'HTTP/1.1 302 Moved Temporarily\r\nLocation: https://www.divo.ru/%D0%B4%D0%B0%D0%BD%D0%BD%D1%8B%D0%B9-%D1%80%D0%B5%D1%81%D1%83%D1%80%D1%81-%D0%B7%D0%B0%D0%B1%D0%BB%D0%BE%D0%BA%D0%B8%D1%80%D0%BE%D0%B2%D0%B0%D0%BD/?UrlRedir=http%3A%2F%2Fwww.hizb-ut-tahrir.org%2f\r\nContent-Length: 0\r\nCache-Control: max-age=0, no-cache, no-store, must-revalidate\r\nPragma: no-cache\r\nConnection: close\r\n\r\n',
-      '''\u003chtml\u003e\r\n\u003chead\u003e\r\n\u003ctitle\u003eæ¡æç¶²é é»æ\u003c/title\u003e\r\n\u003cmeta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\"\u003e\r\n\u003cMETA HTTP-EQUIV=\"PRAGMA\" CONTENT=\"NO-CACHE\"\u003e\r\n\u003cmeta name=\"viewport\" content=\"initial-scale=1.0\"\u003e\r\n\u003cstyle\u003e\r\n  #content {\r\n    border:3px solid#aaa;\r\n    background-color:#fff;\r\n    margin:1.5em;\r\n    padding:1.5em;\r\n    font-family:Tahoma,Helvetica,Arial,sans-serif;\r\n    font-size:1em;\r\n  }\r\n  h1 {\r\n    font-size:1.3em;\r\n    font-weight:bold;\r\n    color:#196390;\r\n  }\r\n  b {\r\n    font-weight:normal;\r\n    color:#196390;\r\n  }\r\n\u003c/style\u003e\r\n\u003c/head\u003e\r\n\u003cbody bgcolor=\"#e7e8e9\"\u003e\r\n\u003cdiv id=\"content\"\u003e\r\n\u003ch1\u003eéç¾è³è¨å®å¨é²è­· - æ¡æç¶²é é»æ\u003c/h1\u003e\r\n\u003cp\u003eå¦ææ¨çå°è©²ç«é¢è³è¨ï¼è¡¨ç¤ºæ¨è¢«å¤æ·å­åéæ­£å¸¸è¡çºç¶²ç« \u003cspan style=\"color:red;\"\u003e(æ¡æç¶²ç«)\u003c/span\u003e\u003c/p\u003e\r\n\u003cp\u003eè³è¨èª²å·²å°æ­¤ç¶²é é»æï¼å¦æç¢ºå®è©²ç¶²é æ¯è¢«èª¤å¤è«è¯ç¹«: éç¾è³è¨èª²-ç³»çµ±ç¶­éçµï¼è¬è¬ã\u003c/p\u003e\r\n\u003cp\u003e\u003cb\u003eä½¿ç¨è:\u003c/b\u003e 141.212.123.175 \u003c/p\u003e\r\n\u003cp\u003e\u003cb\u003eç¶²å:\u003c/b\u003e rtyutgyhefdafioasfjhjhi.com/ \u003c/p\u003e\r\n\u003cp\u003e\u003cb\u003eåé¡:\u003c/b\u003e command-and-control \u003c/p\u003e\r\n\u003c/div\u003e\r\n\u003c/body\u003e\r\n\u003c/html\u003e\r\n'''
-    ]
-    # yapf: enable
-
-    expected_matches = [
-        False,
-        True,
-        None,
-        True,
-        True,
-        True,
-    ]
-    expected_signatures = [
-        'x_generic_akamai_error', 'a_prod_drweb_3', None, 'c_isp_ru_iterika',
-        'c_isp_ru_divo_2', 'e_unk_style_red'
-    ]
-
-    match_pairs = [matcher.match_page(page) for page in pages]
-    matches, signatures = zip(*match_pairs)
-    self.assertListEqual(list(matches), expected_matches)
-    self.assertListEqual(list(signatures), expected_signatures)
-
-  def test_flattenmeasurement_dns(self) -> None:
-    """Test flattening of Satellite measurements."""
-
-    filenames = [
-        'gs://firehook-scans/satellite/CP_Satellite-2020-09-02-12-00-01/interference.json',
-        'gs://firehook-scans/satellite/CP_Satellite-2020-09-02-12-00-01/interference.json',
-        'gs://firehook-scans/satellite/CP_Satellite-2020-09-02-12-00-01/interference.json',
-        'gs://firehook-scans/satellite/CP_Satellite-2021-03-01-12-00-01/interference.json'
-    ]
-
-    # yapf: disable
-    interference = [
-        """{
-          "resolver":"67.69.184.215",
-          "query":"asana.com",
-          "answers":{
-            "151.101.1.184":["ip", "http", "cert", "asnum", "asname"],
-            "151.101.129.184":["ip", "http", "cert", "asnum", "asname"],
-            "151.101.193.184":["ip", "http", "cert", "asnum", "asname"],
-            "151.101.65.184":["ip", "cert", "asnum", "asname"]
-          },
-          "passed":true
-        }
-        """,
-        """{
-          "resolver":"145.239.6.50",
-          "query":"www.ecequality.org",
-          "answers":{
-            "160.153.136.3":[]
-          },
-          "passed":false
-        }
-        """,
-        """{
-          "resolver":"185.228.168.149",
-          "query":"www.sportsinteraction.com",
-          "error":"no_answer"
-        }
-        """,
-        """{
-          "vp":"114.114.114.110",
-          "location":{
-            "country_name":"China",
-            "country_code":"CN"
-          },
-          "test_url":"abs-cbn.com",
-          "response":{
-            "104.20.161.135":[],
-            "104.20.161.134":[],
-            "rcode":["0","0","0"]
-          },
-          "passed_control":true,
-          "connect_error":false,
-          "in_control_group":true,
-          "anomaly":true,
-          "confidence":{
-            "average":0,
-            "matches":[0,0],
-            "untagged_controls":false,
-            "untagged_response":false
-          },
-          "start_time":"2021-03-05 15:32:05.324807502 -0500 EST m=+6.810936646",
-          "end_time":"2021-03-05 15:32:05.366104911 -0500 EST m=+6.852233636"
-        }
-        """
-    ]
-
-    expected = [
-      {
-        'domain': 'asana.com',
-        'ip': '67.69.184.215',
-        'date': '2020-09-02',
-        'error': None,
-        'blocked': False,
-        'success': True,
-        'received': {'ip': '151.101.1.184', 'matches_control': 'ip http cert asnum asname'},
-        'measurement_id': ''
-      },
-      {
-        'domain': 'asana.com',
-        'ip': '67.69.184.215',
-        'date': '2020-09-02',
-        'error': None,
-        'blocked': False,
-        'success': True,
-        'received': {'ip': '151.101.129.184', 'matches_control': 'ip http cert asnum asname'},
-        'measurement_id': ''
-      },
-      {
-        'domain': 'asana.com',
-        'ip': '67.69.184.215',
-        'date': '2020-09-02',
-        'error': None,
-        'blocked': False,
-        'success': True,
-        'received': {'ip': '151.101.193.184', 'matches_control': 'ip http cert asnum asname'},
-        'measurement_id': ''
-      },
-      {
-        'domain': 'asana.com',
-        'ip': '67.69.184.215',
-        'date': '2020-09-02',
-        'error': None,
-        'blocked': False,
-        'success': True,
-        'received': {'ip': '151.101.65.184', 'matches_control': 'ip cert asnum asname'},
-        'measurement_id': ''
-      },
-      {
-        'domain': 'www.ecequality.org',
-        'ip': '145.239.6.50',
-        'date': '2020-09-02',
-        'error': None,
-        'blocked': True,
-        'success': True,
-        'received': {'ip': '160.153.136.3', 'matches_control': ''},
-        'measurement_id': ''
-      },
-      {
-        'domain':  'www.sportsinteraction.com',
-        'ip': '185.228.168.149',
-        'date': '2020-09-02',
-        'error': "no_answer",
-        'blocked': None,
-        'success': False,
-        'received': None,
-        'measurement_id': ''
-      },
-      {
-        'domain': 'abs-cbn.com',
-        'ip': '114.114.114.110',
-        'country': 'CN',
-        'date': '2021-03-05',
-        'start_time': '2021-03-05 15:32:05.324807502 -0500 EST m=+6.810936646',
-        'end_time': '2021-03-05 15:32:05.366104911 -0500 EST m=+6.852233636',
-        'error': None,
-        'blocked': True,
-        'success': True,
-        'received': {'ip': '104.20.161.135', 'matches_control': ''},
-        'rcode': ["0", "0", "0"],
-        'measurement_id': ''
-      },
-      {
-        'domain': 'abs-cbn.com',
-        'ip': '114.114.114.110',
-        'country': 'CN',
-        'date': '2021-03-05',
-        'start_time': '2021-03-05 15:32:05.324807502 -0500 EST m=+6.810936646',
-        'end_time': '2021-03-05 15:32:05.366104911 -0500 EST m=+6.852233636',
-        'error': None,
-        'blocked': True,
-        'success': True,
-        'received': {'ip': '104.20.161.134', 'matches_control': ''},
-        'rcode': ["0", "0", "0"],
-        'measurement_id': ''
-      }
-    ]
-    # yapf: enable
-
-    result = []
-    for filename, i in zip(filenames, interference):
-      flattener = beam_tables.FlattenMeasurement()
-      flattener.setup()
-      rows = flattener.process((filename, i))
-      # remove random measurement id
-      for row in rows:
-        row['measurement_id'] = ''
-        result.append(row.copy())
-    self.assertListEqual(result, expected)
 
   def test_read_satellite_tags(self) -> None:
     """Test reading rows from Satellite tag files."""
@@ -1265,7 +642,7 @@ class PipelineMainTest(unittest.TestCase):
     # yapf: enable
 
     # mock data for the global interference IP - DOMAIN mapping
-    beam_tables.INTERFERENCE_IPDOMAIN = {
+    flatten.INTERFERENCE_IPDOMAIN = {
         '104.20.161.134': {'abs-cbn.com', 'xyz.com', 'blah.com'},
         '198.35.26.96': {'ar.m.wikipedia.org'},
     }
